+++ linera-service/tests/local_net_tests.rs
@@ -29,7 +29,7 @@ use linera_service::cli_wrappers::local_net::Database;
 use linera_service::cli_wrappers::local_net::get_node_port;
 use linera_service::cli_wrappers::local_net::ProcessInbox;
 use linera_service::cli_wrappers::local_net::PathProvider;
-use common::INTEGRATION_TEST_GUARD;
+use common::{get_fungible_account_owner, INTEGRATION_TEST_GUARD};
 
 
 /// Clears the `RUSTFLAGS` environment variable, if it was configured to make warnings fail as
@@ -498,4 +498,58 @@ async fn test_storage_service_linera_net_up_simple() -> Result<()> {
         }
     }
     panic!("Unexpected EOF for stderr");
+}
+
+#[cfg(feature = "benchmark")]
+#[cfg(any(
+    feature = "dynamodb",
+    feature = "scylladb",
+    feature = "storage-service"
+))]
+#[cfg_attr(feature = "storage-service", test_case(LocalNetConfig::new_test(Database::Service, Network::Grpc) ; "storage_service_grpc"))]
+#[cfg_attr(feature = "storage-service", test_case(LocalNetConfig::new_test(Database::Service, Network::Tcp) ; "storage_service_tcp"))]
+#[cfg_attr(feature = "scylladb", test_case(LocalNetConfig::new_test(Database::ScyllaDb, Network::Grpc) ; "scylladb_grpc"))]
+#[cfg_attr(feature = "dynamodb", test_case(LocalNetConfig::new_test(Database::DynamoDb, Network::Grpc) ; "aws_grpc"))]
+#[cfg_attr(feature = "scylladb", test_case(LocalNetConfig::new_test(Database::ScyllaDb, Network::Tcp) ; "scylladb_tcp"))]
+#[cfg_attr(feature = "dynamodb", test_case(LocalNetConfig::new_test(Database::DynamoDb, Network::Tcp) ; "aws_tcp"))]
+#[test_log::test(tokio::test)]
+async fn test_end_to_end_benchmark(mut config: LocalNetConfig) -> Result<()> {
+    use std::collections::BTreeMap;
+
+    use fungible::{FungibleTokenAbi, InitialState, Parameters};
+
+    config.num_other_initial_chains = 2;
+    let _guard = INTEGRATION_TEST_GUARD.lock().await;
+    tracing::info!("Starting test {}", test_name!());
+
+    let (mut net, client) = config.instantiate().await?;
+
+    assert_eq!(client.load_wallet()?.num_chains(), 2);
+    // Launch local benchmark using all user chains and creating additional ones.
+    client.benchmark(2, 4, 10, None).await?;
+    assert_eq!(client.load_wallet()?.num_chains(), 4);
+
+    // Now we run the benchmark again, with the fungible token application instead of the
+    // native token.
+    let account_owner = get_fungible_account_owner(&client);
+    let accounts = BTreeMap::from([(account_owner, Amount::from_tokens(1_000_000))]);
+    let state = InitialState { accounts };
+    let (contract, service) = client.build_example("fungible").await?;
+    let params = Parameters::new("FUN");
+    let application_id = client
+        .publish_and_create::<FungibleTokenAbi, Parameters, InitialState>(
+            contract,
+            service,
+            &params,
+            &state,
+            &[],
+            None,
+        )
+        .await?;
+    client.benchmark(2, 5, 10, Some(application_id)).await?;
+
+    net.ensure_is_running().await?;
+    net.terminate().await?;
+
+    Ok(())
 }
